rename wklib to webknossos
