+++ README.md
@@ -8,5 +8,5 @@ Collection of libraries around webKnossos, please see the respective subfolders:
 
 Python library for creating and working with [webKnossos](https://webknossos.org) [WKW](https://github.com/scalableminds/webknossos-wrap) datasets. WKW is a container format for efficiently storing large, scale 3D image data as found in (electron) microscopy.
 
-## [webKnossos-lib (wklib)](wklib)
+## [webKnossos](webknossos) Python Library
 :construction: WIP
