+++ webknossos/testdata/complex_property_ds/color/1/header.wkw
