+++ webknossos/testdata/complex_property_ds/datasource-properties.json
