+++ webknossos/testdata/complex_property_ds/segmentation/1/header.wkw
