+++ webknossos/testdata/old_wk_dataset/color/1/header.wkw
