+++ webknossos/testdata/old_wk_dataset/color/2/header.wkw
