+++ webknossos/testdata/old_wk_dataset/datasource-properties.json
