+++ webknossos/testdata/simple_wk_dataset/color/1/header.wkw
