+++ webknossos/testdata/simple_wk_dataset/color/1/z0/y0/x0.wkw
