+++ webknossos/testdata/simple_wk_dataset/datasource-properties.json
