+++ webknossos/webknossos/dataset/compress_utils.py
