+++ webknossos/webknossos/dataset/defaults.py
