+++ webknossos/webknossos/geometry/bounding_box.py
