+++ webknossos/webknossos/geometry/mag.py
