+++ webknossos/webknossos/utils.py
